+++ src/main/kotlin/com/fightclub/fight_club_server/match/controller/MatchSocketController.kt
@@ -5,9 +5,7 @@ import com.fightclub.fight_club_server.common.exception.SocketCodeException
 import com.fightclub.fight_club_server.match.constants.MatchSocketSuccessCode
 import com.fightclub.fight_club_server.match.dto.DeclineRequest
 import com.fightclub.fight_club_server.match.dto.ReadyRequest
-import com.fightclub.fight_club_server.match.exception.MatchNotFoundException
-import com.fightclub.fight_club_server.match.exception.UserIsNotParticipantException
-import com.fightclub.fight_club_server.match.service.MatchSocketService
+import com.fightclub.fight_club_server.match.service.MatchService
 import com.fightclub.fight_club_server.user.exception.UserNotFoundException
 import com.fightclub.fight_club_server.user.repository.UserRepository
 import org.springframework.messaging.handler.annotation.MessageMapping
@@ -18,7 +16,7 @@ import java.security.Principal
 
 @Controller
 class MatchSocketController(
-    private val matchSocketService: MatchSocketService,
+    private val matchService: MatchService,
     private val userRepository: UserRepository,
     private val messagingTemplate: SimpMessagingTemplate
 ) {
@@ -27,7 +25,7 @@ class MatchSocketController(
     fun ready(@Payload readyRequest: ReadyRequest, principal: Principal) {
         try {
             val user = userRepository.findByEmail(principal.name) ?: throw UserNotFoundException()
-            val data = matchSocketService.updateReadyStatus(user, readyRequest)
+            val data = matchService.updateReadyStatus(user, readyRequest)
             messagingTemplate.convertAndSend(
                 "/ws/sub/match/room/${readyRequest.matchId}",
                 SocketResponse.success(MatchSocketSuccessCode.MATCH_READY_UPDATED_SUCCESS, data)
@@ -46,7 +44,7 @@ class MatchSocketController(
     fun decline(@Payload declineRequest: DeclineRequest, principal: Principal) {
         try {
             val user = userRepository.findByEmail(principal.name) ?: throw UserNotFoundException()
-            val data = matchSocketService.declineMatch(user, declineRequest)
+            val data = matchService.declineMatch(user, declineRequest)
             messagingTemplate.convertAndSend(
                 "/ws/sub/match/room/${declineRequest.matchId}",
                 SocketResponse.success(MatchSocketSuccessCode.DECLINE_MATCH_SUCCESS, data)
