+++ src/main/kotlin/com/fightclub/fight_club_server/match/domain/ChatMessageType.kt
@@ -1,0 +1,5 @@
+package com.fightclub.fight_club_server.match.domain
+
+enum class ChatMessageType {
+    TEXT, IMAGE
+}
