+++ src/main/kotlin/com/fightclub/fight_club_server/match/repository/ChatMessageRepository.kt
@@ -1,0 +1,6 @@
+package com.fightclub.fight_club_server.match.repository
+
+import com.fightclub.fight_club_server.match.domain.ChatMessage
+import org.springframework.data.jpa.repository.JpaRepository
+
+interface ChatMessageRepository: JpaRepository<ChatMessage, Long>
