--- src/main/kotlin/com/fightclub/fight_club_server/match/repository/MatchMessageRepository.kt
@@ -1,6 +1,0 @@
-package com.fightclub.fight_club_server.match.repository
-
-import com.fightclub.fight_club_server.match.domain.MatchMessage
-import org.springframework.data.jpa.repository.JpaRepository
-
-interface MatchMessageRepository: JpaRepository<MatchMessage, Long>
