+++ src/main/kotlin/com/fightclub/fight_club_server/match/service/MatchService.kt
@@ -1,20 +1,20 @@
 package com.fightclub.fight_club_server.match.service
 
-import com.fightclub.fight_club_server.match.constants.MatchConstants
-import com.fightclub.fight_club_server.match.dto.MatchInfoResponse
-import com.fightclub.fight_club_server.match.dto.MatchResponse
+import com.fightclub.fight_club_server.match.domain.MatchStatus
+import com.fightclub.fight_club_server.match.dto.*
 import com.fightclub.fight_club_server.match.exception.MatchNotFoundException
 import com.fightclub.fight_club_server.match.exception.UserIsNotParticipantException
 import com.fightclub.fight_club_server.match.mapper.MatchMapper
-import com.fightclub.fight_club_server.match.repository.MatchMessageRepository
+import com.fightclub.fight_club_server.match.repository.ChatMessageRepository
 import com.fightclub.fight_club_server.match.repository.MatchRepository
 import com.fightclub.fight_club_server.user.domain.User
+import jakarta.transaction.Transactional
 import org.springframework.stereotype.Service
 
 @Service
 class MatchService(
     private val matchRepository: MatchRepository,
-    private val matchMessageRepository: MatchMessageRepository,
+    private val chatMessageRepository: ChatMessageRepository,
     private val matchMapper: MatchMapper,
 ) {
 
@@ -32,6 +32,38 @@ class MatchService(
         }
 
         return matchMapper.toInfoResponse(match, user)
+    }
+
+    @Transactional
+    fun updateReadyStatus(user: User, readyRequest: ReadyRequest): MatchReadyResponse {
+        val match = matchRepository.findByIdWithLock(readyRequest.matchId).orElseThrow { throw MatchNotFoundException() }
+
+        if (!match.isParticipant(user)) {
+            throw UserIsNotParticipantException()
+        }
+
+        match.updateReadyStatus(user)
+        matchRepository.save(match)
+
+        return matchMapper.toMatchReadyResponse(match, user)
+    }
+
+    @Transactional
+    fun declineMatch(user: User, declineRequest: DeclineRequest): DeclineMatchResponse {
+        val match = matchRepository.findByIdWithLock(declineRequest.matchId).orElseThrow { throw MatchNotFoundException() }
+
+        if (!match.isParticipant(user)) {
+            throw UserIsNotParticipantException()
+        }
+
+        if (match.status == MatchStatus.DECLINED) {
+            throw MatchNotFoundException()
+        }
+
+        match.declineMatch()
+        matchRepository.save(match)
+
+        return matchMapper.toDeclineMatchResponse(match, user)
     }
 
     fun startMatchStreaming(matchId: Long, user: User): Unit {
